Refactor error components into a single directory
