+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/App.js
@@ -17,8 +17,8 @@
 import AppLayout from "./AppLayout";
 import Cells from "./pages/cells";
 import {ColorProvider} from "./pages/common/color";
-import ErrorBoundary from "./pages/common/ErrorBoundary";
-import NotFound from "./pages/common/NotFound";
+import ErrorBoundary from "./pages/common/error/ErrorBoundary";
+import NotFound from "./pages/common/error/NotFound";
 import Overview from "./pages/overview";
 import React from "react";
 import SignIn from "./pages/SignIn";
