+++ system/control-plane/global/components/observability/org.wso2.vick.observability.ui/src/pages/common/error/NotFound.js
@@ -1,19 +1,17 @@
 /*
  * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
  * http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 import ErrorOutline from "@material-ui/icons/ErrorOutline";
@@ -52,7 +50,7 @@ const NotFound = (props) => (
         <div className={props.classes.notFound}>
             <ErrorOutline className={props.classes.notFoundContentIndicator}/>
             <div className={props.classes.notFoundContent}>
-                {props.content ? props.content : "Unable to Find What You were Looking For"}
+                {props.message ? props.message : "Unable to Find What You were Looking For"}
             </div>
         </div>
     </div>
@@ -60,7 +58,7 @@ const NotFound = (props) => (
 
 NotFound.propTypes = {
     classes: PropTypes.object.isRequired,
-    content: PropTypes.string
+    message: PropTypes.string
 };
 
 export default withStyles(styles, {withTheme: true})(NotFound);
